Break up large test file
